requirements.txt
Flask
flask-cors
numpy
Pillow
tensorflow
